bugfix: Diagram.singleton pi/po maps wrong target
Codomain of `f.G.pi` and `f.G.po` for `f = Diagram.singleton()` was set
to an int.
This causes assertions to fail when composing with other diagrams having
different values for this codomain, since coproducts are not defined in
that case.
Fix: use None as codomain (meaning pi/po are maps with values in the
natural numbers)

Also:

- Added `many_singletons` hypothesis generator (analogous to `many_diagrams`)
- Add tests for tensor/composition of singleton diagrams