Minor bugfixes + tests
Use `np.fromiter` to build arrays in Array backend.
This prevents bugs when creating finite functions representing
non-finite signatures like the below:

    FiniteFunction(None, [(2,3)], dtype='object')

Previously, this would create an array `[2, 3]` because numpy interprets
it as a nested array.

Added a test for FrobeniusFunctor for a functor which is not
identity-on-objects.
This tests a case fixed in a previous commit.